# Speed up hashing by using bigger chunks