Turns out, 1KB chunks are _really_ inefficient.

64KB appears to work pretty well. Tested two packages on my laptop:
- 1.8GB, 50170 files: 40s -> 16s
- 11GB, 1383 files: 180s -> 28s

Chunks bigger than 64KB don't make much difference.